requirements.txt
pymupdf
pytesseract
Pillow
python-docx
